Create a shared library, plugable to clang.
Handle all occurrances of diff(f, args) and treat them as an entry point for
our differentiator. Essentially we collect all calls to diff and extract the 
argument out (in the example f) and try to differentiate it and dump the differentiated
functions's AST.

Of course the for now we return the function to be differentiated.