Implement differentiation of x*x applying the chain rule.
Now we traverse the AST for multiplicative binary operators and apply the chained
rule on those. For example x*x*x will become x*(x*x)' + x'*x*x.

Currently there are very few 'interesting' nodes, i.e nodes that we would like to
transform. For the rest of the nodes we need to just 'deep-copy' (aka clone) them
in the new function body.
 
The cloning happens with the StmtClone, which is LLVM licensed AST node cloner,
that I found in the Scout project of TU-Dresden. It is changed to fit our needs.

Our custom visitor produces tiny objects, NodeContext, where we could put more
visitation context if we need to.


* Fix testing case to verify the particular case. Remove redundant CHECK-s.
* Move the rest of the examples in the other test. We need better restructuring.