Remove the useless components that we link in.
This brought a static const to the data section of libclad.so. This lead to a
discrepancy, because the implementation of llvm::APFloat uses the address of the
const static object. If the operation was executed outside libclad.so it'd return
different address from if it was requested within libclad.so. Thus we hit an
assert in clang.