Add function calls to reverse mode, fix bugs
Now we can differentiate function calls in reverse mode, in a similar manner as it is done in forward mode.
For functions with a single parameter, we look for a `f_darg0` function in `custom_derivatives` namespace,
for more parameters we look for `f_grad`.

Now reverse mode should support everything that was also implemented in forward mode.

Some bugs were also fixed.

Fixes #40.