Reimplement runtime check.
When clad works as a clang plugin usually we see the runtime quite early
because #includes usually go first. In the case of interpreters such as
cling the clad runtime can come at any point in time.

This patch activates the plugin only when we include
clad/Differentiator/Differentiator.h'. We speed up cases where we do not
have code to be processed by the plugin.