Implement an option CLAD_BUILD_STATIC_ONLY.
CLAD_BUILD_STATIC_ONLY instructs cmake to build only static libraries.
This is useful when we would like to link clad to an existing shared
library.

This patch gives us a way to build clad on platforms where
LLVM_ENABLE_PLUGINS is off (this is the case for Windows and CYGWIN).

Fixes #78.