Implement reverse pass, store forward pass subexpressions, reduce cloning complexity
This commit implements several things:

1. Implement proper forward and reverse pass in the reverse mode
..* The forward pass does all the computations of the original function
and stores them. Reverse pass computes the derivatives using stored
values from the forward pass. To solve the issue with local variables
and scopes, forward pass stores intermediate values in variables
declared in the beginning of derivative function's scope.
..* While separation into forward and reverse pass is not required when
there are no control-flow dependent variable reassignments, it is
crucial if we want to support reassignments (not implemented yet).
..* Since C++ functions may have several control-flow based points of
return, this requires special treatment. When forward pass contains
`return` statement that is not the last statement in the pass,
reverse computation should not start in the beginning of the reverse
pass, but at some further point. Previously, this was implemented by
cloning the corresponding "tail" of the reverse pass and putting that
tail immediately after the point of return. This is not efficient, as
it requires to clone the reverse pass once per each `return` statement
in function's body. Now, we only have a singe forward and reverse pass,
and preliminary return from a forward pass is implemented as a `goto`
statement to a corresponding label in the reverse pass.

2. Store intermediate expression results in temporary variables in the
forward pass of the reverse mode
..* As it was done before for the forward mode, intermediate results of
expressions of the forward pass are now stored in intermediate
variables, which are later reused in both forward and reverse pass.
This allows to avoid recomputing repeated expressions and cloning long
expression trees. Previosly, in the reverse mode this was done only for
"reverse" part, now it is also enabled in the forward pass.

3. Reduce cloning complexity in the reverse pass
..* Now the complexity of building the gradient function via the
reverse mode should be linear instead of quadratic, as the issue of
repeated cloning of expression trees is solved.