Implement variable reassignments in the reverse mode
Now the reverse mode supports variable reassignments, potentially depending on control-flow (e.g. IfStmts).
At the moment it is assumed that each assignment statement is executed at most once (i.e. it is not inside a loop/backwards-pointing goto), as
otherwise it is reqired to store result of the assignment once per each execution (e.g. in a stack).