Optimize stores, fix issues with compound statements in reverse mode
Several optimizations for `StoreAndRef` and `GlobalStoreAndRef` are implemented.
In `StoreAndRef`, useless temporaries are no longer created for expressions like `+x`, `-x` and array subscripts.
In `GlobalStoreAndRef`, constant literals are no longer stored. This will be useful when implementing for-loops to avoid stores into stack.

Additionally, in the reverse mode, an issue when visiting conditional operators is fixed:
In C++, in conditional operators like `(c ? a : b)` only `a` or `b` is executed, depending on `c`. We did not account for this when visiting, therefore
side effects of both `a` and `b` could be applied, independent of `c`. Now, an IfStmt is created and only one side will be evaluated.