[CI] Run clad with clang which has asserts enabled.
This gives us a great deal of coverage. Clang has well-developed assert system
which often can detect malformed AST. This is very essential for clad as until
today I would need to run that by hand, locally on every pull request.

At the moment we rely on the fact that llvmlab has such clang builds but
unfortunately only for OSX.