Run all clang-9 builds on ubuntu 16.04.
Apparently a recent ubuntu 18 update made clang-9 produce incompatible code.

In that particular build setup we download clang-9 to compile libclad.so which
is then passed as a plugin to various clang versions. For example,
`clang-{8,9,10,11,12 -load libclad.so`.

It looks like that clang-9 stopped producing "compatible" code to that version
of ubuntu 18.