[codecov] Use a secure post-build action to upload the coverage info.
Apparently, the token-less upload has become problematic recently for external
forks. We want to try uploading the coverage information using tokens but that
can only happen outside of the PR jobs as there might be malicious PRs which can
steal tokens.

This patch relies on the post-pr build action to upload the relevant information.