Fix building of nested name specifiers in getArg function
This commit modifies code in getArg function to build correct and
complete nested name specifiers for the call operator.
It also fixes the failing asserts in clang debug build on running tests for
differentiating functors. Credit goes to @grimmmyshini and @sudo-panda for
bringing this error to attention.

The root cause of the failing asserts was lambda types not having any
associated identifiers. This was causing issues while creating correct
nested name specifiers for the call operator.