Add array diff support to ReverseModeVisitor
The signature of the grad function has changed and it creates as many diff params as the number of requested args by the user. The diff params are placed at the end and in the same order as the original args.

The diff params are of the type clad::array_ref<return_type> where return_type is the return type of the original function. This change was introduced as clad needs to know the size of the arrays passed to the derived function when calling a function inside the derived function.

Since the number of output params depend on the number of requested args clad is unable to know the number of output params during type deduction. To solve this an overload for the grad function is created that has the same number of output params as determined at type deduction stage (equal to the number of params of the original function). The extra args are present just to match the DerivedFnType are given default arg functionality by using the pad_and_execute function. The original grad function is then called inside the overload function using the required params.

Eg.
// Requested function:
int f(int a, int* b) {
  . . .
}

// Derived function:
void f_grad1(int a, int* b, int* d_b) {
  . . .
}

// Derived function overload:
void f_grad1(int a, int* b, int* d_b, int* d_0) {
  return f_grad1(a, b, d_b);
}

This has also led to the Derive functions in ForwardModeVisitor, ReverseModeVisitor, JacobianModeVisitor, HessianModeVisitor to return an extra argument called OverloadFunctionDecl. It is a nullptr in the Visitors except in ReverseModeVisitor which passes the overload whenever it is created. Both the overloaded and the actual derived FunctionDecl is returned because both of them need to be registered on the reDecl chain and also the actual diff function needs to be dumped even when the overload is created.