Fix segfault on clad::tape resize.
This fixes the memory bug in clad::tape that causes clad to fail with a segmentation fault on tape resizes due to an invalid write error. This error arose due to the fact that the newly allocated pointer was being incremented directly instead of being saved to assign to the base data pointer later.