Add support for differentiating switch stmt in the reverse mode AD
This commit adds support for differentiating switch statements in the
reverse mode AD.

The basic idea used to differentiate switch statement is that in the
forward pass, processing of the statements of the switch statement body
always starts from a case/default label and ends at a break statement or
at the end of the switch body. Similarly, in the reverse pass, processing
of the differentiated statements of the switch statement body will start
from the statement just above the break statement that was hit or from the
last differentiated statement in the case when no break statement was hit.

Thus, we can keep track of which break statement was hit in the forward pass
or if no break statement was hit at all in a variable. This information is
further used by an auxiliary switch statement in the reverse pass to jump the
execution to the correct point -- that is, differentiated statement of the
statement just before the break statement that was hit in the forward pass.

In this strategy, each switch case statement of the original function gets
transformed to an if condition in the reverse pass. The if condition decides
at runtime whether the processing of the differentiated statements of the
switch statement body should stop or continue. This is again based on the
fact that the processing of statements of the switch statement body always
starts at a case statement.