Add initial support for diff class types in the forw mode
This commit adds initial support for differentiating class types with
respect to scalar numerical types using the forward mode AD.

The forward mode automatic differentiation allows us to compute
derivative of all the output variables with respect to a single
input variable (independent variable) in a single pass. Mathematically,
the forward mode AD allows us to efficiently compute columns of the
jacobian matrix. Up till now, we are not utilising this ability of
the forward mode AD -- since we only supported differentiating scalar
type with respect to a scalar type.
Thus, differentiation of aggregate types in the forward mode AD will
allow us to utilize this ability and obtain multiple derivatives in a
single call to the derived function.

Clad requires the following assumptions to be true when differentiating a class type:

- Class should represent a real vector space.
- Class should have a default constructor. Class variables initialised
  by the default constructor should represent 0 tangent vector -- that
  is, all real data members should be equal to 0.

Differentiation of class types in the forward mode AD currently have the following limitations:

- Calls to member functions are not supported.
- Calls to overloaded operators are not supported.
- Initialising class type variables using non-default constructor are not supported.
- Class cannot have pointer or reference data members.

Most of these limitations will be removed soon.