Add initial support for differentiating user-defined types in the reverse mode.
This commit adds initial support for differentiating a scalar built-in numerical
type with respect to user-defined types using the reverse-mode AD.

The reverse mode automatic differentiation allows to efficiently compute the
derivative of a single output variable with respect to all the input variables.
Mathematically, the reverse mode AD allows to efficiently compute rows of the
Jacobian matrix.

Clad requires the following assumptions to be true while differentiating with
respect to class types in the reverse mode AD:

- Class should represent a real vector space.
- Class should have a default constructor that zero initialises all the data members
  of the class. That is, a newly default constructed class object should represent 0 tangent vector.

Differentiation with respect to class types in the reverse mode AD currently has the following limitations:

- Calls to member functions are not supported
- Calls to overloaded operators are not supported
- Initialising class type variables using non-default constructors that take non-literal arguments (non-zero derivatives)
  are not supported.
- Class cannot have pointer or reference data members.

Most of these limitations will be removed soon.

This commit also modifies the overloaded derivative function mechanism as described in the issue #394. Two most important changes
are as follows:

- Overloaded derived function is now created for all the derived functions that are generated using the reverse mode AD.
  Earlier, we were not creating the overloaded derived function if gradient was requested with respect to all the original
  function parameters.

- Type of all the parameters derivatives is: `clad::array_ref<void>`. It gets converted to (by `static_cast`)
  to correct type before being passed as the argument to the actual derived function call.