Fix diff of ref variable in loops in rev mode
This PR adds two major changes/features to the
`ReverseModeVisitor` design. These changes/features are:

expression designed to be used in the forward sweep of Reverse Mode.

Until now, the simplified behaviour of `Visit*` member functions of
`ReverseModeVisitor` can be described as:

When a `Visit*` function is called with statement `S` as an argument,
it returns a `StmtDiff` object that contains a forward sweep statement
and a derived statement. The forward sweep statement contains the behaviour
of the argument statement `S` along with any other functionality required for
Reverse Mode AD.
The derived statement is a little more complicated. If the argument `S` is an
expression, then the derived statement is the corresponding adjoint of the
expression designed to be used in the reverse sweep. Otherwise, if the argument
`S` is a control flow statement, compound statement -- non-expression statements,
then the derived statement is the corresponding block (compound statement) that
should be inserted in the reverse sweep.

Notice that, the derivative returned by `StmtDiff` is only designed to be used
in the reverse sweep. Also, no other functionality exists that provides derivatives
to be used in the forward sweep.

This PR modifies `StmtDiff` to also store derivatives expressions designed to be
used in the forward sweep of Reverse Mode.
This also involves computation of derivatives for forward sweep in various
`Visit*` functions, most notably `VisitArraySubscriptExpr`.

Practical use case of this functionality:
```cpp
for (int i=0; i < 5; ++i) {
  double& ref = arr[i];
}
```
Forward sweep of the loop should be as follows:
```cpp
// declarations
double* _d_ref = 0;

for (int i=0; i < 5; ++i) {
  _d_ref = &_d_arr[i];
  double& ref = arr[i];
}
```

Please note that the derivative expression of `arr[i]` as returned by
`Visit*` functions is `_d_arr[r0]` where `r0` will be a reverse sweep
temporary variable and thus, the expression, `_d_arr[r0]` cannot be
used in the forward sweep.

This functionality allows to add statements at the top of the loop body
in the reverse sweep. The main motivation behind this is to add declaration
statements just at the beginning of the loop body.

Practical use case of this functionality:

```cpp
for (int i=0; i < 5; ++i) {
  double& ref = arr[i];
  res += ref;
}
```

gradient function:

```cpp
// declarations
double* _d_ref = 0;
clad::tape<double*> t0;

int t1 = 0;
// forward sweep
for (int i=0; i < 5; ++i) {
   ++t1;
  _d_ref = &_d_arr[i];
  clad::push(t0, _d_ref);
  double& ref = arr[i];
  res += ref;
}

// reverse sweep
for (; t1 > 0; --t1) {
  int r0 = clad::pop(t0);  <---------- declarations at top of the loop body
  *r0 += _d_res;
}
```
Closes #440