Patch for the failure of functions with void return type and with return statement
Prior to this patch, clad would throw an assertion error when attempting to differentiate void functions that contain a return statement.
The assertion error would indicate that there was an attempt to differentiate a nullptr.
This Patch checks if the there is an expression associsted with the null pointer and only if there is one, we differentiate it.

Added tests for the same

This patch fixes #432