Detect Request for use of Enzyme as Backend by user
This patch is the first step towards integrating enzyme AD with Clad.
The user can now mention that he/she wants to use enzyme for AD instead
of clad by using the API calls:
	clad::differentiate<clad::opts::use_enzyme>(func,"...");
and	clad::gradient<clad::opts::use_enzyme>(func);

Via this patch one can successfully recognise the request for Enzyme AD
backend. It is implemented via Overloading the function templates for
differentiate and gradient. Currently after successfully recognising a Enzyme
request, an empty function is created to show that we have identified the
request for enzyme. That is,

clad::gradient<clad::opts::use_enzyme>(func);
generates the function:
	void func_grad_enzyme(...function arguments... , ...arguments to store derivatives...){
	}

Currently generation of dummy function for clad::differentiate<clad::opts::use_enzyme>(...) is not implemented.
These functions must be modified in future patches to insert valid code for enzyme to work on.