Generate code for Enzyme autodiff for functions with Pointer/array arguments
For functions of type
	double myfunction(double* arr){...};
OR 	double myfunction(double arr[3]){...};

we generate derivative code of type:
	void d_myfunction(double* arr, clad::array_ref<double> _d_arr){
		double* d_arr = _d_arr.ptr();
		__enzyme_autodiff_myfunction(myfunction, arr, d_arr);
	}

Further on enzyme will handle the function differentiation

This Commit also Downloads and Installs Enzyme in the Clad Directory, if
the flag "-DENABLE_ENZYME_BACKEND=On" is passed to cmake during project configuration.

This commit also contains tests for the above feature in test/Gradient/Enzyme.C

This commit also adds the enzyme pass to the llvm optimisation pipeline so that
it can run at the right moment.