Add the correct literal types arguments potentially used in the pushforward.
The call argument and its differentiated call argument should have the same type.
That is not the case where we synthesize 0 or 1 as an integer literals when
applying the differentiation rules.

This type mismatch is problematic for the analytical derivatives via pushforwards
where we have ValueAndPushforward<T, T> X_pushforward(T x, T d_x). In that case
if we pass arguments of type `double` and an `int` we fail to instantiate and
thus fail to use the relevant pushforward.

This patch adds an implicit cast to the right type if the type of the call
argument differs from the type of the differentiated one.