Add Documentation for Integration of Enzyme with Clad
This Commit adds documentation on how to use Enzyme within Clad.
It conveys to a user on how to enable Enzyme backend within Clad
and use it for Differentiating functions within clad, but with Enzyme.

It also introduces Enzyme, Clad and their approaches to Automatic
Differentiation to a new user.