Add support for Pullback of a Pushforward Function
This commit temporarily adds support for computing the pullback of a
pushforward function.

The commit differentiates an `InitListExpr` that represents the `ValueAndPushforward`
class instance knowing what fields in the class it must use as adjoints for finding
gradient. Though this is a temporary fix, the correct solution must differentiate at
Constructor level rather than at the `InitListExpr` level.

This commit also adds support to Visit a Static Cast Expr as it was needed when
differentiating `log_pushforward`.