Fix incompatible array pushes to tape.
Previously, if we see an array that is passed to a function, we would
try to insert it into the tape. However, this is wrong because best
case, we would end up only cloning the array pointer and not the array
itself. This commit introduces a new tape push function that captures a
snapshot (clones the array contents) when a pointer is pushed to the
tape.

This method is very inefficient given that most functions implicity
treat arrays as read-only data structures, so we end up cloning the
unchanged array several times. One possible fix is to identify if the
function the array is passed to expects a const array and then do not
clone the array.

This one commit, by no means, is a complete solution to the problem with
arrays in loops. For example, for input arrays, we still end up pushing
the pointer instead of the whole array. Moreover, variable sized arrays
with array sizes declared in-function are also not supported.