Remove dependency on LastKnownGoodFiles*.txt
Fixes vgvassilev#523 This commit aims to remove clad's dependecy on LastKnownGoodFiles*.txt as
there is  better version check infrastructure present in clad. This commit also makes
the validation of clang version default behavior. And also adds a new routine to
compare the clang version at runtime and the clang version for which clad was built.