Improve vector initialization for vector forward mode
Improved initialization of vectors while differentiating variable declarations and added test cases to demonstrate vector mode support for programs with loops.

New ways of initializing clad::array is added by passing (size, expression...).
This is done to allow initializing the array by all constant values or expressions which evaluate to a single value;
for ex: clad::array<double> arr(5UL, 1) can be used to initialize a vector of size 5 with all entries as 1.