Remove the redundant store of primal return result into a variable.
At the end of the forward pass for every non-void function, we store the return
value in <function_name>_return. However, we only need to store those variables
that could be used in control flow or be present as arguments in functions or
operators (which the return value never could be). Therefore, storing it is
unnecessary. In this pull request, all the <function_name>_return variables
are removed.

Note: when we need _ret_value in error estimation mode we now directly assign
`_ret_value = <return_expression>` instead of
`_ret_value = <function_name>_return`.
`MultiplexExternalRMVSource::ActBeforeFinalisingVisitReturnStmt()`,
`ErrorEstimationHandler::ActBeforeFinalisingVisitReturnStmt()` and
`ErrorEstimationHandler::SaveReturnExpr()` now take the return expression as
their only argument.

Fixes vgvassilev/clad#582.