Add LLVM17 support.
This patch replaces the artificial handling of the scopes and accesses the
private CurScope member of Sema. This allows us to build precise lexical scope
setup for when building AST nodes. That's essential for lambda support in
clang17 because it tightly couples both Scopes and DeclContexts.

The change in test/FirstDerivative/Loops.C is because the pretty printer did not
print correctly the variables defined in the condition of the loop until clang17.

Fixes #669