[cmake] Use the compiler clad is built for if compiling benchmarks and tests.
We need to attach the clad plugin to the compiler invocations in order to
either benchmark or test clad's infrastructure. The best way I could find is by
setting/resetting the CMAKE_CXX_COMPILER variable which is a hack that works.
Note that we do not have a cmake guarantee that will work but since that's in
the core of our testing infrastructure we will find out quickly if, when and
where it breaks.