[cmake] Do not link libLLVM.so on platforms that enable it by default.
AddLLVM.cmake can be configured in a way where it puts all LLVM libraries into
a single libLLVM.so file. Then we can configure LLVM at build time to explicitly
link libLLVM.so file to all targets that depend on llvm. That is very bad for
plugins because it inserts the symbols once again into the binary causing double
initialization of globals and other problems.

This breaks the clad package for conda.

This issue is related to #715.