Fix condition declarations & assignments, enable logical operators in for loops
These changes fix the differentiation of variable declarations in for
loop conditions that used to result into wrong derivatives. The commit
also tackles the problem of having an assignment operator that affects
the derivative in the for-loop condition, as well as adds support for
logical operators in for-loops and allows to combine assignments with
them.

Fixes: #273