Introduce `CLAD_DISABLE_TESTS` flag to build system
The testing and benchmarking relies on a hack that temporariliy changes
the compiler flags. This doesn't work in all environments, and it would
be nice if the the tests could be disabled so it's possible to build
Clad without this hack.

This makes it possible to build Clad on NixOS.