Keep adjoint types same as the original ones on function global scope.
The purpose of this commit is to merge decl stmts with their initialization points in the reverse mode for pointer- and reference-type variables. All of the simplifications happen when the original statement is used on the function global scope because otherwise, the declaration has to be "promoted to the function global scope". In particular
1) Adjoints of reference-type variables, e.g.
```
double *_d_ref = 0;
double &ref = x;
_d_ref = &_d_x;
```
can be refactored to
```
double &_d_ref = _d_x;
double &ref = x;
```
Notice: there's also no need to change the adjoint type to ``double*`` in this case (like was done previously).

2) Adjoints of pointer-type variables, e.g.
```
double *_d_ptr = 0;
double *ptr = &x;
_d_ptr = &_d_x;
```
can be similarly refactored to
```
double *_d_ptr = &_d_x;
double *ptr = &x;
```

Also, after this commit, adjoints on a function global scope are generated right before the original (cloned) variables and not at the very beginning of the function. Adjoints are never used before the original variables so this change will not affect anything.