Merge "_t" variable declarations with their initialization points on function global scope.
The purpose of this commit is to merge decl stmts (e.g. ``double _t0;``) with their initialization points (e.g.``_t0 = x;``) in the reverse mode whenever this is possible. All of the simplifications happen when the original statement is used on the function global scope because otherwise, the declaration has to be "promoted to the function global scope". The simplification is done for `_t` variables created by `GlobalStoreAndRef`, `DelayedGlobalStoreAndRef`, and `StoreAndRestore`, e.g.
```
double _t0;
...
_t0 = x;
```
can be refactored to
```
...
double _t0 = x;
```
Fixes #525.